refactored non SCHO modules out of the repository and changed key class and method names
