+++ .gitignore
@@ -13,3 +13,4 @@ digits_config.py
 fmnist_config.py
 imdb_config.py
 run.py
+PlotEngine.py
